requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
websockets==12.0
openai==1.12.0
deepl==1.17.0
jinja2==3.1.3
python-dotenv==1.0.1
aiohttp==3.9.3 